Removed RunAwayScript + Fixed UI Scale
Removed the new "RunAwayScript" in favor of the old one: during my testing I found that the Things would run in circles around you or sometimes into the net. I took the new one out and added the old one back in. Attempted to fix UI scale when game is built and run on a machine.